improved structure of files, fixed bug in follow, removed redundant code
