+++ minitwit-api/sim/go.mod
@@ -1,0 +1,4 @@
+module minitwit.com/sim
+
+go 1.21
+
