+++ minitwit-api/sim/sim-facade.go
@@ -1,10 +1,9 @@
-package latest
+package sim
 
 import (
 	"encoding/json"
 	"net/http"
 	"os"
-	"strconv"
 )
 
 func UpdateLatest(r *http.Request) {
@@ -13,19 +12,6 @@ func UpdateLatest(r *http.Request) {
 	if latest != "" {
 		_ = os.WriteFile("./latest_processed_sim_action_id.txt", []byte((latest)), 0644)
 	}
-}
-
-func Get_latest(w http.ResponseWriter, r *http.Request) {
-	content, _ := os.ReadFile("./latest_processed_sim_action_id.txt")
-	latest_command_id, err := strconv.Atoi(string(content))
-	if err != nil {
-		latest_command_id = -1
-	}
-	json.NewEncoder(w).Encode(struct {
-		Latest int `json:"latest"`
-	}{
-		Latest: latest_command_id,
-	})
 }
 
 func Is_authenticated(w http.ResponseWriter, r *http.Request) bool {
